+++ CHANGES.md
@@ -6,6 +6,8 @@ Version 0.5.0
 
 To be released.
 
+ -  Added `-f`/`--stop-on-fail` option.
+
 
 Version 0.4.0
 -------------