+++ CHANGES.md
@@ -9,6 +9,7 @@ To be released.
  -  Added `-f`/`--stop-on-fail` option.
  -  Added `-p`/`--parallel` option.
  -  Added `-H`/`--hang-seconds` option.
+ -  Applied AppDomains.
 
 
 Version 0.4.0