+++ CHANGES.md
@@ -1,6 +1,15 @@
 xUnit.net Unity Runner Changelog
 ================================
 
+
+Version 0.6.0
+-------------
+
+To be released.
+
+ -  Upgraded Unity Editor from 2020.3.4f1 to 2021.3.5f1.
+
+
 Version 0.5.0
 -------------